+++ packages/ckbtc/src/errors/minter.errors.ts
@@ -1,5 +1,6 @@
-import { nonNullish } from "@dfinity/utils";
+import { fromNullable, nonNullish } from "@dfinity/utils";
 import type {
+  PendingUtxo,
   RetrieveBtcError,
   RetrieveBtcWithApprovalError,
   UpdateBalanceError,
@@ -10,7 +11,21 @@ export class MinterTemporaryUnavailableError extends MinterGenericError {}
 export class MinterAlreadyProcessingError extends MinterGenericError {}
 
 export class MinterUpdateBalanceError extends MinterGenericError {}
-export class MinterNoNewUtxosError extends MinterUpdateBalanceError {}
+export class MinterNoNewUtxosError extends MinterUpdateBalanceError {
+  readonly pendingUtxos: PendingUtxo[];
+  readonly requiredConfirmations: number;
+  constructor({
+    pending_utxos,
+    required_confirmations,
+  }: {
+    pending_utxos: [] | [PendingUtxo[]];
+    required_confirmations: number;
+  }) {
+    super();
+    this.pendingUtxos = fromNullable(pending_utxos) || [];
+    this.requiredConfirmations = required_confirmations;
+  }
+}
 
 export class MinterRetrieveBtcError extends MinterGenericError {}
 export class MinterMalformedAddressError extends MinterRetrieveBtcError {}
@@ -49,7 +64,7 @@ export const createUpdateBalanceError = (
   }
 
   if ("NoNewUtxos" in Err) {
-    return new MinterNoNewUtxosError();
+    return new MinterNoNewUtxosError(Err.NoNewUtxos);
   }
 
   // Handle types added in the backend but not yet added in the frontend